buscar antes de eliminar y editar grupo

+++ src/modules/groups/application/deleteCase.js
@@ -1,3 +1,5 @@
+const notFound = require('../domain/exceptions/groupNotFound')
+
 module.exports = class {
 
     constructor (authGpsRepository) {
@@ -5,8 +7,13 @@ module.exports = class {
     }
 
     execute (idGroup) {
+
         return  new Promise((resolve, reject) => {
-            this.repository.delete(idGroup).then( destroy => resolve(destroy)).catch(reject)
-        });
+            this.repository.findById(idGroup).then((find)=>{
+                if(find === null) return reject(new notFound(idGroup))
+                this.repository.delete(idGroup).then( destroy => resolve(destroy)).catch(reject)
+            }).catch(reject)
+
+        })
     }
 }
